fix: unsettled promise when request is aborted and compression is used and maxRedirects is set to 0
The axios request is left in an unsettled state (waits forever) when using compression
and maxRedirects is set to 0 and the request gets aborted after some chunks are sent.

The problem is that the response `abort` event is not being emitted and the request state
inside the `responseStream.on('error')` callback is `destroyed: true, aborted: false`.

I see that previoysly axios was checking `req.aborted` instead of `req.destroyed` inside
the `responseStream.on('error')` callback, but since node deprecated the `req.aborted`, it was
changed to `req.destroyed` in #4787.

We could change that back to `req.aborted`, but I think checking `rejected` would be
a better option, so we can be sure that we are catching all the errors
(in case there are some more weird cases).